feat: improve data loading
Move loading data to the abstract class. This will reduce the complexity in the actual report classes. And the majority of the reports are either json or yaml so no need to define this each time on a report level.
